* We are now fully C++ compliant. Yay! ... damn that was a pain... * Added Recast and Detour code to cgame for now for starting on moving to navmeshes.

+++ codemp/game/bg_weapons.h
@@ -163,7 +163,7 @@ typedef struct weaponData_s
 	int		dmg;
 	int		dmgAlt;
 
-	float	boltSpeed;
+	int		boltSpeed;
 	int		accuracy;
 
 	int		splashDmg;
